moved Forms under SearchPage and added a way to get to results page
