+++ src/views/SearchPage/SearchPage.scss
@@ -79,3 +79,11 @@
     justify-content: center;
   }
 }
+
+.submit-block {
+  text-align: center;
+  padding: 2em 0;
+  .btn-submit {
+    margin: 0 auto;
+  }
+}
